requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
docopt==0.6.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.1.1
frozenlist==1.6.0
fsspec==2025.5.1
future==1.0.0
idna==3.10
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
line-bot-sdk==3.17.1
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai-whisper==20240930
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.31.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.25.1
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
sympy==1.14.0
tiktoken==0.9.0
tinycss2==1.4.0
torch==2.7.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
yarg==0.1.9
yarl==1.20.0

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
coverage==7.3.2
